// src/types.ts
export type EssayType = 
  | 'ta-cay-coi' 
  | 'ke-chuyen-da-doc-da-nghe' 
  | 'cam-xuc-nhan-vat-van-hoc' 
  | 'thuat-lai-su-viec' 
  | 'neu-y-kien-hien-tuong'
  | 'neu-y-kien-cau-chuyen'
  | 'cam-xuc-nguoi-than'
  | 'viet-thu'
  | 'viet-don'
  | 'ta-con-vat'
  | 'viet-huong-dan-bao-cao';